Add favorite recipes tab and being reformatting elements
