+++ src/Components/Tile/recipewindow.css
@@ -31,6 +31,16 @@
     float: right;
 }
 
+.recipe-details {
+    vertical-align: "top";
+    /*padding: 50px;*/
+}
+
+.recipe-image {
+    align-items: end;
+    border: "4px solid #999";
+}
+
 .steps{
     /*color: rgb(0, 0, 0);*/
     font-size: 16px;
